IntrusiveLinearRegionMap used in copies

+++ src/null/NullDebug.cpp
@@ -4,7 +4,7 @@ void nanos_wait_for_full_initialization(void)
 {
 }
 
-long int nanos_get_num_cpus()
+unsigned int nanos_get_num_cpus()
 {
 	return 1;
 }
